railway-fastapi-torch-macos122/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
Pillow
requests
torchvision